RV32 ISS implement debug_interface

+++ vp/src/core/rv32/syscall_if.h
@@ -13,8 +13,8 @@ struct iss_syscall_if {
 
 	virtual unsigned get_syscall_register_index() = 0;
 
-	virtual uint32_t read_register(unsigned idx) = 0;
-	virtual void write_register(unsigned idx, uint32_t value) = 0;
+	virtual uint64_t read_register(unsigned idx) = 0;
+	virtual void write_register(unsigned idx, uint64_t value) = 0;
 
 	// virtual uint32_t get_hart_id() = 0;
 };
